refactor(app) : 패키지 변경
- 연관 model 그룹화
