+++ src/main/java/model/Team.java
@@ -2,13 +2,13 @@ package model;
 
 public enum Team {
     RED("한"),
-    BLUE("초")
-    ;
+    BLUE("초");
     private final String team;
 
     Team(String team) {
         this.team = team;
     }
+
     public String getTeam() {
         return team;
     }
