+++ src/main/java/model/janggi_board/JanggiBoardSetUp.java
@@ -1,9 +1,18 @@
-package model;
+package model.janggi_board;
 
 import static model.Team.BLUE;
 import static model.Team.RED;
 
 import java.util.Map;
+import model.Point;
+import model.piece.Byeong;
+import model.piece.Cha;
+import model.piece.Jang;
+import model.piece.Ma;
+import model.piece.Pho;
+import model.piece.Piece;
+import model.piece.Sa;
+import model.piece.Sang;
 
 public enum JanggiBoardSetUp {
