+++ src/main/java/model/piece/Byeong.java
@@ -1,20 +1,29 @@
-package model;
+package model.piece;
 
 import java.util.List;
 import java.util.Map;
 import java.util.stream.IntStream;
+import model.Path;
+import model.Point;
+import model.Team;
 
-public class Sa extends Piece {
+public class Byeong extends Piece {
 
-    public Sa(Team team) {
+    public Byeong(Team team) {
         super(team);
-        pieceName = PieceName.SA;
+        pieceName = PieceName.BYEONG;
     }
 
     @Override
     public boolean isValidPoint(Point beforePoint, Point targetPoint) {
-        List<Integer> horizontal = List.of(0, 0, -1, 1);
-        List<Integer> vertical = List.of(1, -1, 0, 0);
+        int moveForward = 1;
+
+        if (team == Team.RED) {
+            moveForward = -1;
+        }
+
+        List<Integer> horizontal = List.of(0, -1, 1);
+        List<Integer> vertical = List.of(moveForward, 0, 0);
 
         return IntStream.range(0, horizontal.size())
                 .anyMatch(i -> horizontal.get(i) + beforePoint.x() == targetPoint.x()
@@ -35,4 +44,5 @@ public class Sa extends Piece {
         }
         return true;
     }
+
 }
