+++ src/main/java/model/piece/Jang.java
@@ -1,8 +1,11 @@
-package model;
+package model.piece;
 
 import java.util.List;
 import java.util.Map;
 import java.util.stream.IntStream;
+import model.Path;
+import model.Point;
+import model.Team;
 
 public class Jang extends Piece {
