+++ src/main/java/model/piece/Ma.java
@@ -1,22 +1,27 @@
-package model;
+package model.piece;
 
 import java.util.List;
 import java.util.Map;
 import java.util.stream.IntStream;
+import model.Path;
+import model.Point;
+import model.Team;
 
-public class Sang extends Piece {
-    public Sang(Team team) {
+public class Ma extends Piece {
+
+    public Ma(Team team) {
         super(team);
-        pieceName = PieceName.SANG;
+        pieceName = PieceName.MA;
     }
 
     @Override
     public boolean isValidPoint(Point beforePoint, Point targetPoint) {
-        List<Integer> horizontal = List.of(-2, 2, 3, 3, 2, -2, -3, -3);
-        List<Integer> vertical = List.of(3, 3, 2, -2, -3, -3, -2, 2);
+        List<Integer> horizontal = List.of(-1, 1, 2, 2, 1, -1, -2, -2);
+        List<Integer> vertical = List.of(2, 2, 1, -1, -2, -2, -1, 1);
 
         return IntStream.range(0, horizontal.size())
-                .anyMatch(i -> horizontal.get(i) + beforePoint.x() == targetPoint.x() && vertical.get(i) + beforePoint.y() == targetPoint.y());
+                .anyMatch(i -> horizontal.get(i) + beforePoint.x() == targetPoint.x()
+                        && vertical.get(i) + beforePoint.y() == targetPoint.y());
     }
 
     @Override
@@ -27,20 +32,18 @@ public class Sang extends Piece {
         int unitVectorX = vectorX / Math.abs(vectorX);
         int unitVectorY = vectorY / Math.abs(vectorY);
 
-        Point middlePoint1 = new Point(targetPoint.x() - unitVectorX, targetPoint.y() - unitVectorY);
-        Point middlePoint2 = new Point(targetPoint.x() - unitVectorX * 2, targetPoint.y() - unitVectorY * 2);
+        Point middlePoint = new Point(targetPoint.x() - unitVectorX, targetPoint.y() - unitVectorY);
         Point endPoint = new Point(targetPoint.x(), targetPoint.y());
 
         Path path = new Path();
-        path.addPoint(middlePoint1);
-        path.addPoint(middlePoint2);
+        path.addPoint(middlePoint);
         path.addPoint(endPoint);
         return path;
     }
 
     @Override
-    public boolean canMove(Map<Piece,Boolean> piecesOnPathWithTargetOrNot) {
-        if (piecesOnPathWithTargetOrNot.size() >= 2) {
+    public boolean canMove(Map<Piece, Boolean> piecesOnPathWithTargetOrNot) {
+        if (piecesOnPathWithTargetOrNot.size() == 2) {
             return false;
         }
         if (piecesOnPathWithTargetOrNot.size() == 1) {
@@ -51,5 +54,4 @@ public class Sang extends Piece {
         }
         return true;
     }
-
 }
