+++ src/main/java/model/piece/Sang.java
@@ -1,20 +1,22 @@
-package model;
+package model.piece;
 
 import java.util.List;
 import java.util.Map;
 import java.util.stream.IntStream;
+import model.Path;
+import model.Point;
+import model.Team;
 
-public class Ma extends Piece {
-
-    public Ma(Team team) {
+public class Sang extends Piece {
+    public Sang(Team team) {
         super(team);
-        pieceName = PieceName.MA;
+        pieceName = PieceName.SANG;
     }
 
     @Override
     public boolean isValidPoint(Point beforePoint, Point targetPoint) {
-        List<Integer> horizontal = List.of(-1, 1, 2, 2, 1, -1, -2, -2);
-        List<Integer> vertical = List.of(2, 2, 1, -1, -2, -2, -1, 1);
+        List<Integer> horizontal = List.of(-2, 2, 3, 3, 2, -2, -3, -3);
+        List<Integer> vertical = List.of(3, 3, 2, -2, -3, -3, -2, 2);
 
         return IntStream.range(0, horizontal.size())
                 .anyMatch(i -> horizontal.get(i) + beforePoint.x() == targetPoint.x()
@@ -23,29 +25,26 @@ public class Ma extends Piece {
 
     @Override
     public Path calculatePath(Point beforePoint, Point targetPoint) {
-        // 1. 이동 벡터를 구한다
-        // 2. 이동 단위 벡터를 구한다.
-        // 3. 단위 벡터 역벡터를 도착점에 더한다.
-        // 3.이랑 도착점을 Path에 담아서 리턴
-
         int vectorX = targetPoint.x() - beforePoint.x();
         int vectorY = targetPoint.y() - beforePoint.y();
 
         int unitVectorX = vectorX / Math.abs(vectorX);
         int unitVectorY = vectorY / Math.abs(vectorY);
 
-        Point middlePoint = new Point(targetPoint.x() - unitVectorX, targetPoint.y() - unitVectorY);
+        Point middlePoint1 = new Point(targetPoint.x() - unitVectorX, targetPoint.y() - unitVectorY);
+        Point middlePoint2 = new Point(targetPoint.x() - unitVectorX * 2, targetPoint.y() - unitVectorY * 2);
         Point endPoint = new Point(targetPoint.x(), targetPoint.y());
 
         Path path = new Path();
-        path.addPoint(middlePoint);
+        path.addPoint(middlePoint1);
+        path.addPoint(middlePoint2);
         path.addPoint(endPoint);
         return path;
     }
 
     @Override
     public boolean canMove(Map<Piece, Boolean> piecesOnPathWithTargetOrNot) {
-        if (piecesOnPathWithTargetOrNot.size() == 2) {
+        if (piecesOnPathWithTargetOrNot.size() >= 2) {
             return false;
         }
         if (piecesOnPathWithTargetOrNot.size() == 1) {
@@ -56,4 +55,5 @@ public class Ma extends Piece {
         }
         return true;
     }
+
 }
