+++ src/main/java/view/InputView.java
@@ -3,7 +3,7 @@ package view;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
-import model.JanggiBoard;
+import model.janggi_board.JanggiBoard;
 import model.Point;
 import model.Team;
 
@@ -12,10 +12,10 @@ public class InputView {
         System.out.printf("%s 나라의 차례\n", team.getTeam());
         System.out.println("이동할 기물의 위치를 선택하세요.");
         List<Integer> beforePointInput = Arrays.stream(sc.nextLine().split(" ")).map(Integer::parseInt).toList();
-        Point beforePoint = Point.of(beforePointInput.get(0), beforePointInput.get(1));
+        Point beforePoint = Point.of(beforePointInput.get(0)-1, beforePointInput.get(1)-1);
         System.out.println("이동될 위치를 선택하세요.");
         List<Integer> targetPointInput = Arrays.stream(sc.nextLine().split(" ")).map(Integer::parseInt).toList();
-        Point targetPoint = Point.of(targetPointInput.get(0), targetPointInput.get(1));
+        Point targetPoint = Point.of(targetPointInput.get(0)-1, targetPointInput.get(1)-1);
         janggiBoard.move(beforePoint, targetPoint);
     }
 }
