+++ src/test/java/model/DotTest.java
@@ -2,6 +2,9 @@ package model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import model.janggi_board.Dot;
+import model.piece.Jang;
+import model.piece.Piece;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
