+++ src/test/java/model/JanggiBoardTest.java
@@ -1,8 +1,9 @@
 package model;
 
-import static model.JanggiBoardSetUp.INNER_ELEPHANT;
+import static model.janggi_board.JanggiBoardSetUp.INNER_ELEPHANT;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import model.janggi_board.JanggiBoard;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,7 +12,7 @@ public class JanggiBoardTest {
     @DisplayName("장기말 세팅 테스트 - 초기 장기말 갯수 확인")
    void  test1() {
         JanggiBoard janggiBoard = new JanggiBoard(INNER_ELEPHANT);
-        
+
         assertThat(janggiBoard.countPiece()).isEqualTo(32);
     }
 }
